spearate commnad logic into modules
