+++ rootstock/commands/__init__.py
@@ -1,0 +1,18 @@
+"""Command modules for the Rootstock CLI."""
+
+from .init import cmd_init
+from .install import cmd_install
+from .manifest import cmd_manifest
+from .resolve import cmd_resolve
+from .serve import cmd_serve
+from .status import cmd_list, cmd_status
+
+__all__ = [
+    "cmd_init",
+    "cmd_install",
+    "cmd_list",
+    "cmd_manifest",
+    "cmd_resolve",
+    "cmd_serve",
+    "cmd_status",
+]
